Some small fixes to the issue templates forms (#70)
I've removed the shell rendering from the issue templates for custom and
documentation issues as it doesn't really feel right to default to that
form for those types of issues. Whilst doing so I also fixed an issue
with inconsistent e.g./ex. for the placeholder text and added examples
for linking to SRS tickets.